bmsui: Limit statdisplay refresh rate
Contents
The refresh rate of a statdisplay is limited to a minimum of 250ms.

Motivation
If a value on a statdisplay is borderline, it can rapidly oscillate between the two states.
This can make it difficult to read the value.
In the future when more program memory is available averaging should be implemented.