sensors.cpp: Comment about current sensor ADC reference voltage
Contents
A comment is added to reinforce that the current sensor inputs on the MCU use VDD as an ADC reference, not 2.5V.

Motivation
Since '2.5V' is still in the code as a placeholder for that parameter, it can be confusing what the reference is on first glance.
If the reference was actually 2.5V, the current sensors wouldn't work.